requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
Jinja2
python-telegram-bot>=22.0.0
aiohttp>=3.8.0
scrapy>=2.13.0
twisted>=22.10.0
itemadapter>=0.7.0